+++ src/util.h
@@ -163,13 +163,13 @@ bool should_copy_mmap_region(const std::string& filename,
  * Return an fd referring to a new shmem segment with descriptive
  * |name| of size |num_bytes|.
  */
-ScopedFd create_shmem_segment(const std::string& name, size_t num_bytes);
+ScopedFd create_shmem_segment(const std::string& name, uint64_t num_bytes);
 
 /**
  * Ensure that the shmem segment referred to by |fd| has exactly the
  * size |num_bytes|.
  */
-void resize_shmem_segment(ScopedFd& fd, size_t num_bytes);
+void resize_shmem_segment(ScopedFd& fd, uint64_t num_bytes);
 
 enum cpuid_requests {
   CPUID_GETVENDORSTRING,
